// DocumentConverterFunction/src/generation/type/enum/internal.document.sub.type.ts
export enum InternalDocumentSubType {
  /**
   * Касовий чек (реалізація)
   */
  SALE = 0,

  /**
   * Видатковий чек (повернення)
   */
  RETURN = 1,

  /**
   * Чек операції «службове внесення»/«отримання авансу»
   */
  SERVICE_DEPOSIT = 2,

  /**
   * Чек операції «отримання підкріплення»
   */
  ADDITIONAL_DEPOSIT = 3,

  /**
   * Чек операції «службова видача»/«інкасація»
   */
  SERVICE_ISSUE = 4,

  /**
   * Чек сторнування попереднього чека
   */
  CANCEL = 5,
}